fixes for tasks
fix potential deadlock:

* if we are decrementing task count another thread that might read task
count just after it was decremented might decide that the task list is
empty and delete it before we had a chance to complete the cleanup code
that is part of decrementing the task counter.

we fix this by adding a protection flag which gets incremented at the
start of the protected section and decremented at the end - we then add
this protection flag to the number of flags as it symbolises at
task-in-flight.

this saves us from using a mutex, which would be another way of solving
this issue.

* rename tasks_count -> task_count

* remove blocking on task list

if we're blocking on a task list, we can have a situation where a thread
worker resumes to an empty channel and decides to go to sleep. because
the other thread has gone to sleep too as it blocked on its task list,
we both threads are blocked now.