fix possible deadlock when destroying channels
if threads were not signalled to stop before the signal for the channel
to wake up was set, threads would wake once and would not notice that
they were supposed to break out of the inner loop.

requesting thread stop, then nudging the sleeping flags gives threads a
chance to gracefully exit their inner loop.